+++ src/server/tcp.rs
@@ -218,7 +218,7 @@ mod tests {
         impl Service for DummyService {
             type Request = Request<'static>;
             type Response = Response;
-            type Exception = Exception;
+            type Exception = ExceptionCode;
             type Future = future::Ready<Result<Self::Response, Self::Exception>>;
 
             fn call(&self, _: Self::Request) -> Self::Future {
@@ -253,8 +253,8 @@ mod tests {
         impl Service for DummyService {
             type Request = Request<'static>;
             type Response = Response;
-            type Exception = Exception;
-            type Future = future::Ready<Result<Self::Response, Exception>>;
+            type Exception = ExceptionCode;
+            type Future = future::Ready<Result<Self::Response, ExceptionCode>>;
 
             fn call(&self, _: Self::Request) -> Self::Future {
                 future::ready(Ok(self.response.clone()))