+++ CHANGELOG.md
@@ -5,12 +5,13 @@
 
 ## v0.15.0 (Unreleased)
 
-- Added `Exception::new`.
+- Added `Exception::Custom` and `Exception::new`.
 
 ### Breaking Changes
 
 - Added `Exception::Custom`.
 - Removed `TryFrom<u8>` and `#[repr(u8)]` for `Exception`.
+- Renamed `Exception` to `ExceptionCode` to be more consistent with `FunctionCode`.
 
 ## v0.14.1 (2024-09-10)